backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
pydantic
python-dotenv
Werkzeug
gunicorn
psycopg2-binary
Authlib
requests